\name{default_parcube_options}
\alias{default_parcube_options}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
parCube initialization parameters
}
\description{
a LIST of the default options for serial_parCube method
}
\usage{
default_parcube_options()
}
\value{
a LIST of the default options for serial_parCube method
}
\author{
Abdelmoneim Amer Desouki
}
