\name{umls_tnsr}
\docType{data}
\alias{umls_tnsr}

\encoding{utf8}

\title{RDF tensor of UMLS small graph}

\description{
  The dataset is a list that contains UMLS dataset basic graph used by Nickel et al 2011 to 
  test \code{\link{rescal}} .
}

\usage{data(umls_tnsr)}

\format{slot X contains the frontal slices and SO contains the names of the entities 
and P contains the names of the predicates}

\references{
-Maximilian Nickel, Volker Tresp, Hans-Peter-Kriegel,
      "A Three-Way Model for Collective Learning on Multi-Relational Data",
      ICML 2011, Bellevue, WA, USA
}

\keyword{datasets}
\keyword{UMLS}